+++ app/server/api/auth_api.py
@@ -3,10 +3,11 @@ from flask.views import MethodView
 from server import db, sentry
 # from server import limiter
 from server.constants import DENOMINATION_DICT
-from server.models.transfer import TransferAccount
-from server.models.user import User
+from server.models.transfer_account import TransferAccount
 from phonenumbers.phonenumberutil import NumberParseException
-from server.models.models import Organisation, BlacklistToken, EmailWhitelist, CurrencyConversion, TransferUsage
+from server.models.user import User
+from server.models.organisation import Organisation
+from server.models.models import BlacklistToken, EmailWhitelist, CurrencyConversion, TransferUsage
 from server.utils.intercom import create_intercom_android_secret
 from server.utils.auth import requires_auth, tfa_logic
 from server.utils.access_control import AccessControl
