+++ app/server/api/organisation_api.py
@@ -2,9 +2,10 @@ from flask import Blueprint, request, make_response, jsonify
 from flask.views import MethodView
 
 from server import db
-from server.models.models import paginate_query, Organisation, Token
+from server.models.organisation import Organisation
+from server.models.models import paginate_query, Token
 from server.models.user import User
-from server.models.transfer import TransferAccount
+from server.models.transfer_account import TransferAccount
 from server.schemas import organisation_schema, organisations_schema
 from server.utils.auth import requires_auth, show_all
