+++ app/server/api/transfer_account_api.py
@@ -3,7 +3,7 @@ from flask.views import MethodView
 
 from server import db
 from server.models.models import paginate_query
-from server.models.transfer import TransferAccount
+from server.models.transfer_account import TransferAccount
 from server.schemas import transfer_accounts_schema, transfer_account_schema, \
     view_transfer_account_schema, view_transfer_accounts_schema
 from server.utils.auth import requires_auth
