+++ app/server/me_api/credit_transfer.py
@@ -13,7 +13,9 @@ from server.exceptions import (
     InsufficientBalanceError
 )
 from server.models.user import User
-from server.models.transfer import TransferAccount, TransferCard, CreditTransfer
+from server.models.transfer_card import TransferCard
+from server.models.transfer_account import TransferAccount
+from server.models.credit_transfer import CreditTransfer
 from server.models.models import paginate_query
 from server.schemas import me_credit_transfers_schema, me_credit_transfer_schema
 from server.utils.auth import requires_auth, show_all
