+++ test/conftest.py
@@ -28,7 +28,7 @@ def create_blockchain_token(test_client, init_database):
 
 @pytest.fixture(scope='module')
 def create_organisation(test_client, init_database, create_blockchain_token):
-    from server.models.models import Organisation
+    from server.models.organisation import Organisation
     organisation = Organisation(name='Sempo', token=create_blockchain_token)
     db.session.add(organisation)
     db.session.commit()
@@ -107,7 +107,7 @@ def create_user_with_existing_transfer_account(test_client, init_database, creat
 
 @pytest.fixture(scope='module')
 def new_transfer_account():
-    from server.models.transfer import TransferAccount
+    from server.models.transfer_account import TransferAccount
     return TransferAccount()
 
 @pytest.fixture(scope='module')
@@ -124,7 +124,7 @@ def new_disbursement(create_transfer_account_user):
 
 @pytest.fixture(scope='function')
 def new_credit_transfer(create_transfer_account_user, create_blockchain_token):
-    from server.models.transfer import CreditTransfer
+    from server.models.credit_transfer import CreditTransfer
     credit_transfer = CreditTransfer(
         amount=100,
         token=create_blockchain_token,
